requirements.txt
# Core - Python 3.12 compatible versions
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0

# Data Collection
yfinance>=0.2.28
tweepy>=4.14.0
praw>=7.7.1
requests>=2.31.0
websockets>=9.0,<11  # Compatible with alpaca-trade-api
alpaca-trade-api>=3.0.2
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
pymongo>=4.5.0
redis>=5.0.0

# API & Web
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
gunicorn>=21.2.0
eventlet>=0.33.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Monitoring
prometheus-client>=0.18.0
mlflow>=2.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
schedule>=1.2.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

